MNIST - Multiclass Classification With CNN (Keras)
In this notebook, we will build Multi-class classification for the MNIST dataset using Convolution Neural Networks (CNN).
The MNIST dataset is an image dataset of handwritten digits. It has has 60,000 training images and 10,000 test images, each of which are grayscale 28 x 28 sized images.It is a good beginner's dataset to try learning techniques and pattern recognition methods on real-world data while spending minimal efforts on preprocessing and formatting.

In this notebook, we will develop a Convolutional Neural Network using the Keras library to classify these handwritten images.